Extract node types to separate file
